refactor: move function
Move addAuthToConfig function to upm-config-io module
